UpdateJulia

Documentation for UpdateJulia.

UpdateJulia.update_julia

UpdateJulia.update_julia — Function

update_julia(version::AbstractString="")

Install the latest version of Julia from https://julialang.org

If version is provided, installs the latest version that starts with version. If version == "nightly", then installs the bleeding-edge nightly version.

Keyword Arguments

Behavior flags

set_default = (v == latest()) make 'julia' point to installed version.

dry_run = false skip the actual download and instillation

verbose = dry_run print the final value of all arguments

Destination

systemwide = true install for all users, false only installs for current user.

install_location = systemwide ? "/opt" or "/opt/hostedtoolcache/julia/1.7.0" : "/home/runner/.local" directory to put installed binaries

bin = systemwide ? "/usr/local/bin" : "/home/runner/.local/bin" directory to store links to the binaries

Source

os_str = "linux" string representation of the operating system: "linux", "mac", "winnt", or "freebsd".

arch = "x86_64" string representation of the CPU architecture: "x86_64", "i686", "aarch64", "armv7l", or "powerpc64le".

v = ... the VersionNumber to install

url = ... URL to download that version from, if you explicitly set url, also explicitly set v lest they differ